+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Change Log
+
+## Master
+### Fixed
+* Posts without `wp:post_date` or `wp:post_date_gmt` now get defaulted to `Date.today` instead of throwing an error [#8](https://github.com/contentful/wordpress-exporter.rb/issues/8)
+
+
 ## 0.0.2
 ### Fixed
 * Issues with `attachment_description`
+
 
 ## 0.0.1
 ### Other